+++ src/lib/pages/validator-details/components/__stories__/VotingPowerChart.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { VotingPowerChart } from "../VotingPowerChart";
+import { VotingPowerChart } from "../bonded-token-changes/VotingPowerChart";
 
 const meta: Meta<typeof VotingPowerChart> = {
   component: VotingPowerChart,